Assignment/SampleData.cs
using System;
using System.IO;
using System.Linq;
using System.Collections.Generic;

namespace Assignment
{
    public class SampleData : ISampleData
    {
        private string? _FileName;

        public SampleData(string? fileName)
        {
            if (string.IsNullOrEmpty(fileName))
                throw new ArgumentNullException(nameof(fileName));

            _FileName = fileName;
        }

        // 1.
        public IEnumerable<string> CsvRows
        {
            get
            {
                using (var sr = new StreamReader(@_FileName))
                {
                    yield return sr.ReadLine();
                }
            }
        }

        // 2.
        public IEnumerable<string> GetUniqueSortedListOfStatesGivenCsvRows() 
            => throw new NotImplementedException();

        // 3.
        public string GetAggregateSortedListOfStatesUsingCsvRows()
            => throw new NotImplementedException();

        // 4.
        public IEnumerable<IPerson> People => throw new NotImplementedException();

        // 5.
        public IEnumerable<(string FirstName, string LastName)> FilterByEmailAddress(
            Predicate<string> filter) => throw new NotImplementedException();

        // 6.
        public string GetAggregateListOfStatesGivenPeopleCollection(
            IEnumerable<IPerson> people) => throw new NotImplementedException();
    }
}
